o/i/apparmorprompting: renamed promptrequests package to requestprompts
